refactor cards into separate components
